moved visit(all args) from Operation abstract class into specific visitors
It seemed more appropriate to move this behavior into the visitors. It is the action of visiting an expression that should trigger the acceptance of each argument of this expression. Until now, it was the acceptance of a visitor that instantly triggered the acceptance of the visitor by each argument which was a skipping of the visiting action.

In summary : trigger the acceptance of children because of the visiting of the expression and not because of the acceptance of that expression. Acceptance only triggers the visiting of the expression